Keep provider-turn apply logic out of the coordinator hot path
The coordinator still carries too much provider-turn loop and finalization ownership in one file. This change makes provider_turn_apply.rs the authoritative home for apply/finalization, task-progress checkpoint helpers, and runtime tool-discovery persistence so turn_coordinator.rs can stay closer to orchestration instead of re-embedding the same shell logic.

Constraint: Preserve existing provider-turn finalization behavior and checkpoint semantics
Rejected: Start with the riskier provider-turn state-machine extraction first | the apply/finalization shell was already isolated enough to land safely on its own
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep provider-turn apply/finalization helpers in the submodule; do not regrow duplicate implementations inside turn_coordinator.rs
Tested: ./scripts/cargo-local-toolchain.sh check -p loong-app; ./scripts/cargo-local-toolchain.sh test -p loong-app handle_turn_with_runtime_explicit_skill_activation_prefix_injects_skill_context -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong-app provider_turn_request_terminal_phase_builds_inline_provider_error_reply -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong-app waiting_task_progress_record_uses_waiting_status_and_approval_gate_handle -- --test-threads=1; ./scripts/cargo-local-toolchain.sh clippy -p loong-app --all-targets -- -D warnings; ./scripts/cargo-local-toolchain.sh fmt --all -- --check
Not-tested: full workspace test suite; provider-turn runtime state-machine extraction not yet landed in this commit
